code refactoring in login screen
